Added an unnamespaced class map autoloader
This class loader can be used to alleviate problems where switching from a situation of using multiple class loaders and include paths to
using OPL Autoloaders. The format of the class map file is the same as for the regular ClassMapLoader:
array(
	'foo' => 'path/to/foo.php',
	'bar' => 'path/to/bar.php',
)
This of course needs to be serialized and put into a file as well.